PythonProject/pyproject.toml
[project]
name = "aeroopt"
version = "0.0.1"
requires-python = ">=3.9"
dependencies = [
  "pyyaml>=6.0",
  "pandas>=2.0",
  "numpy>=1.24",
  "aerosandbox>=4.0",   # whatever version you use
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
